+++ CHANGELOG.md
@@ -7,11 +7,12 @@
 * **New extension: Sortable.js**, [see here](https://buefy.github.io/#/extensions/sortablejs)
 * **Update Bulma to 0.6.2**
 * **Add Font Awesome 5 support** (http://fontawesome.com/)
+* Add ``columns`` prop to table, you can now pass an array instead of adding scoped slots (for simple tables)
 * Add ``rounded`` prop to Input, Select, Autocomplete, Datepicker and Timepicker, Pagination
 * Add "is-toggle-rounded" type to Tabs
 * Add ``open`` prop to collapse ``trigger`` slot
 * Add ``native`` prop to upload
-* #222 Add ``open-on-focus`` prop to autocomplete (good to simulate a <select> element)
+* #222 Add ``open-on-focus`` prop to autocomplete (good to simulate a select element)
 * #482 Add ``allow-new`` prop to taginput
 * #483 Add ``unselectable-days-of-week`` prop to datepicker
 * #487 Add ``icon-size`` prop to message